make simple_dependency_resolver available more broadly
